network: fix integer comparisons

+++ src/lxc/network.c
@@ -2518,7 +2518,7 @@ static int ifa_get_local_ip(int family, struct nlmsghdr *msg, void **res)
 			 * the address length is correct, but check here just in
 			 * case.
 			 */
-			if (RTA_PAYLOAD(rta) != addrlen)
+			if (RTA_PAYLOAD(rta) != (unsigned int)addrlen)
 				return -1;
 
 			/* We might have found an IFA_ADDRESS before, which we
@@ -2621,7 +2621,7 @@ static int ip_addr_get(int family, int ifindex, void **res)
 				return ret_errno(EINVAL);
 
 			ifa = (struct ifaddrmsg *)NLMSG_DATA(msg);
-			if (ifa->ifa_index == ifindex) {
+			if (ifa->ifa_index == (__u32)ifindex) {
 				if (ifa_get_local_ip(family, msg, res) < 0)
 					return ret_errno(EINVAL);
